runtime-sdk: move gas back into core
