+++ runtime-sdk/src/module.rs
@@ -12,7 +12,7 @@ use crate::{
     storage::Store,
     types::{
         message::MessageEvent,
-        transaction::{CallResult, Transaction, UnverifiedTransaction},
+        transaction::{Call, CallResult, Transaction, UnverifiedTransaction},
     },
 };
 
@@ -165,6 +165,15 @@ pub trait AuthHandler {
         // Default implementation doesn't do any checks.
         Ok(())
     }
+
+    /// Perform any action after authentication, within the transaction context.
+    fn before_handle_call(
+        _ctx: &mut TxContext<'_, '_>,
+        _call: &Call,
+    ) -> Result<(), modules::core::Error> {
+        // Default implementation doesn't do anything.
+        Ok(())
+    }
 }
 
 #[impl_for_tuples(30)]
@@ -182,6 +191,14 @@ impl AuthHandler for Tuple {
         tx: &Transaction,
     ) -> Result<(), modules::core::Error> {
         for_tuples!( #( Tuple::authenticate_tx(ctx, tx)?; )* );
+        Ok(())
+    }
+
+    fn before_handle_call(
+        ctx: &mut TxContext<'_, '_>,
+        call: &Call,
+    ) -> Result<(), modules::core::Error> {
+        for_tuples!( #( Tuple::before_handle_call(ctx, call)?; )* );
         Ok(())
     }
 }
